Fiber#join: save Fiber.current before the callback
Fiber.current was being called in the block passed to Fiber#link. This
was being evaluated in the event loop fiber, causing us to immediately
run the event loop when there are no watchers and thus crash the program
by the next time we transfer into the loop because Loop#run has finished
executing.

This also let me remove that highly suspect `loop do...end` in Loop#run.